Fixing small code mistakes in user and app
